feat(core-app-api): Added entity links to createApp and the BackstageApp
This adds mapping between entity refs (defaulting to groups) and EntityLinks. This is useful to e.g. add group slack channels, and to show them on the entity overview page, as well as the plugin error boundary, to direct people to get help from (and report bugs to) the right team.

Signed-off-by: Gustaf Räntilä <g.rantila@gmail.com>